LayoutTests/accessibility/combobox/combobox-active-element-selected-children-expected.txt

Tests that for a combobox, ActiveElement and SelectedChildren return the same object.

PASS: axCombobox.activeElement === null
PASS: axCombobox.selectedChildren().length === 0
Setting activedescendant to 1 and selected to 2:
notification: AXActiveElementChanged
    activeElement: item1
    selectedChildren: [ item1 ]
notification: AXSelectedChildrenChanged
    activeElement: item1
    selectedChildren: [ item1 ]
Setting activedescendant to 2:
notification: AXActiveElementChanged
    activeElement: item2
    selectedChildren: [ item2 ]
Setting activedescendant to 3:
notification: AXActiveElementChanged
    activeElement: item3
    selectedChildren: [ item3 ]
Selecting 3:
notification: AXSelectedChildrenChanged
    activeElement: item3
    selectedChildren: [ item3 ]

PASS successfullyParsed is true

TEST COMPLETE
